Remove yarn.lock since its unmaintained
Currently no need to check it in and update it, since we don't actively
maintain this project. yarn install will generate a new one locally with
the latest version which will hopefully still work.

Leaving in an unmaintained yarn.lock is a potential security risk for
someone using yarn install --pure-lockfile and will most importantly
just annoy the hell out of everyone because of dependabot.